+++ src/main/java/com/kr1s1s/subtlyd/data/loot_table/BlockLootSD.java
@@ -17,5 +17,7 @@ public class BlockLootSD extends FabricBlockLootTableProvider {
         dropSelf(BlocksSD.SNOW_BRICKS);
         dropSelf(BlocksSD.SNOW_BRICK_STAIRS);
         dropSelf(BlocksSD.SNOW_BRICK_SLAB);
+        dropSelf(BlocksSD.GRASS_SLAB);
+        dropSelf(BlocksSD.DIRT_SLAB);
     }
 }