+++ src/main/java/com/kr1s1s/subtlyd/world/block/BlocksSD.java
@@ -13,6 +13,7 @@ public class BlocksSD {
     public static final Block SNOW_BRICKS = Blocks.register(resourceKey("snow_bricks"), BlockBehaviour.Properties.of().mapColor(MapColor.SNOW).instrument(NoteBlockInstrument.BASEDRUM).strength(1.0F, 0.5F).sound(SoundTypeSD.SNOW_BRICKS));
     public static final Block SNOW_BRICK_STAIRS = registerStair(resourceKey("snow_brick_stairs"), SNOW_BRICKS);
     public static final Block SNOW_BRICK_SLAB = registerSlab(resourceKey("snow_brick_slab"), SNOW_BRICKS);
+    public static final Block GRASS_SLAB = registerSlab(resourceKey("grass_slab"), Blocks.GRASS_BLOCK);
     public static final Block DIRT_SLAB = registerSlab(resourceKey("dirt_slab"), Blocks.DIRT);
 
     public static void init() {}